@@ -1,6 +1,6 @@
 module github.com/intel/amber-client/go-client
 
-go 1.19
+go 1.20
 
 require (
 	github.com/golang-jwt/jwt/v4 v4.4.2
