@@ -7,6 +7,7 @@ package client
 
 import (
 	"crypto/tls"
+	"io"
 	"net/http"
 
 	"github.com/pkg/errors"
@@ -47,7 +48,7 @@ func doRequest(tlsCfg *tls.Config,
 
 	var resp *http.Response
 	if resp, err = client.Do(req); err != nil {
-		return errors.Wrapf(err, "Request to %q failed", req.URL)
+		return errors.Errorf("Request to %q failed: %s", req.URL, err)
 	}
 
 	if resp != nil {
@@ -60,7 +61,11 @@ func doRequest(tlsCfg *tls.Config,
 	}
 
 	if resp.StatusCode != http.StatusOK || resp.ContentLength == 0 {
-		return errors.Errorf("Request to %q failed: StatusCode = %d, ContentLength = %d", req.URL, resp.StatusCode, resp.ContentLength)
+		response, err := io.ReadAll(resp.Body)
+		if err != nil {
+			return errors.Errorf("Failed to read response body: %s", err)
+		}
+		return errors.Errorf("Request to %q failed: StatusCode = %d, Response = %s", req.URL, resp.StatusCode, string(response))
 	}
 
 	return processResponse(resp)
